Clang-Tidy: specify size_t explicitly.
This is because Clang-Tidy thinks that EXPECT_EQ(0u, something) should
actually be EXPECT_EQ(nullptr, something) even if something return a
size_t value. This is clearly a bug in Clang-Tidy, but it’s easy to go
around it, so we may as well do it.

+++ tests/coverage/coverage.cpp
@@ -45,7 +45,7 @@ TEST(Coverage, printer) {
     libcellml::Printer pc(pm);
 
     size_t error_count = pc.errorCount();
-    EXPECT_EQ(0u, error_count);
+    EXPECT_EQ(size_t(0), error_count);
 }
 
 TEST(Coverage, units) {
