+++ tests/model/component_import.cpp
@@ -71,9 +71,9 @@ TEST(ComponentImport, singleImportA) {
 
     EXPECT_EQ(importedComponent->getImportSource(), imp);
 
-    EXPECT_EQ(0u, m.componentCount());
+    EXPECT_EQ(size_t(0), m.componentCount());
     m.addComponent(importedComponent);
-    EXPECT_EQ(1u, m.componentCount());
+    EXPECT_EQ(size_t(1), m.componentCount());
 
     libcellml::Printer printer;
     const std::string a = printer.printModel(m);
@@ -126,9 +126,9 @@ TEST(ComponentImport, nonExistentURLAndParse) {
 
     EXPECT_EQ(importedComponent->getImportSource(), imp);
 
-    EXPECT_EQ(0u, m.componentCount());
+    EXPECT_EQ(size_t(0), m.componentCount());
     m.addComponent(importedComponent);
-    EXPECT_EQ(1u, m.componentCount());
+    EXPECT_EQ(size_t(1), m.componentCount());
 
     libcellml::Printer printer;
     std::string a = printer.printModel(m);
@@ -137,7 +137,7 @@ TEST(ComponentImport, nonExistentURLAndParse) {
     // Parse
     libcellml::Parser parser;
     libcellml::ModelPtr model = parser.parseModel(e);
-    EXPECT_EQ(1u, model->componentCount());
+    EXPECT_EQ(size_t(1), model->componentCount());
     a = printer.printModel(model);
     EXPECT_EQ(e, a);
 }
@@ -194,7 +194,7 @@ TEST(ComponentImport, multipleImportAndParse) {
     // Parse
     libcellml::Parser parser;
     libcellml::ModelPtr model = parser.parseModel(e2);
-    EXPECT_EQ(3u, model->componentCount());
+    EXPECT_EQ(size_t(3), model->componentCount());
     a = printer.printModel(model);
     EXPECT_TRUE((e1 == a) || (e2 == a));
 }
@@ -238,9 +238,9 @@ TEST(ComponentImport, hierarchicalImportAndParse) {
 
     EXPECT_TRUE(i1->isImport());
 
-    EXPECT_EQ(0u, bob->componentCount());
+    EXPECT_EQ(size_t(0), bob->componentCount());
     bob->addComponent(i1);
-    EXPECT_EQ(1u, bob->componentCount());
+    EXPECT_EQ(size_t(1), bob->componentCount());
 
     libcellml::Printer printer;
     std::string a = printer.printModel(m);
@@ -249,7 +249,7 @@ TEST(ComponentImport, hierarchicalImportAndParse) {
     // Parse
     libcellml::Parser parser;
     libcellml::ModelPtr model = parser.parseModel(e);
-    EXPECT_EQ(1u, model->componentCount());
+    EXPECT_EQ(size_t(1), model->componentCount());
     a = printer.printModel(model);
     EXPECT_EQ(e, a);
 }
@@ -313,7 +313,7 @@ TEST(ComponentImport, complexImportAndParse) {
 
     // check component counts
     const libcellml::ComponentPtr constDave = model->getComponent("dave");
-    EXPECT_EQ(1u, constDave->componentCount());
+    EXPECT_EQ(size_t(1), constDave->componentCount());
     const libcellml::ComponentPtr constBob = constDave->getComponent("bob");
-    EXPECT_EQ(2u, constBob->componentCount());
+    EXPECT_EQ(size_t(2), constBob->componentCount());
 }
